Rename project from lightnode > lntools
This renames all projects and readmes to lntools from Lightnode. LNTools is more
fitting name for the goals of the project versus the original project name of
Lightnode. The original project was designed to be a full implementation of the
LN BOLT specs. This rename reflects the purpose of these libraries to provide
tools for various parts of the Lightning Network.