buffer-cursor: readBytes and peakBytes returned shared memory Buffer slices
The slice method of a Buffer returns a new Buffer but points to the same
memory location as the Buffer from which the slice was generated. This
is useful for mutation of the original Buffer but is problematic for
the readBytes and peakBytes use cases that may mutate or change the
results. These two methods now return a new Buffer and copy the data
from the underlying Buffer location instead of using an shared memory
location as returned by slice.