bitcoin: position adjustmenst for indexOfWitness
Originally the indexOfWitness code performed data read operations
when cursing through a transaction. This action results in data
reads to buffers that are not used.

This commit prevents unnecessary buffer reads but adjusting the
BufferCursor position explicility for data that does not need to
be read into memory and analyzed.