wire: direct export of message types
Couple changes to improve type interfance.

1. Message exports are now named object exports instead of direct class
exports. This allows find by reference and go to implementation to work
correctly in editors.

2. Fixed naming convention for certain messages. There was a mix of Message
suffix and no Message suffix on class names. With #1, the addition of the
suffix adds export clarity for the package.

These changes may break modules, however @lntools/wire has not been published
yet, so this is a not a breaking change for deployed libraries.