wire: adding address objects for node-announcment-message refactor
This commit adds polymorphic address types to represent the various
forms of address that can be deserialized from a node-announcment-message.

Additionally, this commit establishes an area for migrating deserialization
code into a separate area for better organization of functional code. The
goal is to move use these functions to build more robust domain objects.