buffer-cursor: update package to 0.1.5
New release allows getting/setting the position of the cursor via
the position property. This feature is useful for seeking through
a buffer by a known amount without reading the underlying data.

For example, this is used in @lntools/bitcoin to calculate tx size
without reading all of the underlying data unnecessarily and
incurring buffer allocation costs.