graph: move message validation to wire
A series of commits have been made to migrate the logic that
formerly resided in the graph project into wire. The net result
was the creation of the GossipFilter class to replace the
GraphManager class in this project.

GraphManager will now take on the sole responsibility of building
graphs from valid messages that are emitted from the GossipFilter
or GossipStore classes.

This will also enable the removal of the Bitcoin and Bitcoind
projects with this package.