wire: creating GossipManager
The GossipManager is an orchestrator for routing gossip traffic
from many different peers. This class handles wiring up the
GossipSyncher and GossipFilter to perform synchronization with
peers and filter incoming messages through a centralized queue.

The net result is that a single subscriber can listen to the
GossipManager for validated routing-gossip traffic.