wire: adding outpoint and mapping of outpoint to scid
When data is received from a chain-client it will often be in the
form of a transaction id (32-byte string) + vout. The short_channel_id
uses a block, tx index, and vout. During validation of a
channel_announcement we obtain the txid and want to establish a link
with the short_channel_id and outpoint so that we can find channels
by either the short_channel_id or the outpoint.

This code is a first pass at adding that functionality to the system.