logger: simplify logger
This change refactors the logger to move away from a central
LogManager that is a singleton and use a single Logger that acts
as the root. All sub-loggers maintain a reference to the root
logger. This root logger configures the transports and sets the
log level. All writes will cascade up to the roots functionality
but retain the area specific information about the sub-logger.