noise: create sub-logger in connect method
Previor to this commit noise used the logger that was supplied in
the connect constructor. Instead, noise will be responsible for
constructing an area specific sub-logger with appropriate instance
name.

This change allows the sub-logger construction to cascade without
requiring calling code to remember to create an appropraite sub-logger.

The major drawback is that we can no longer provide a specific logger
for the module.