wire: adding utxo validation on gossip restore
These change adds UTXO validation for existing channel_announcement
messages that get restored by the GossipManager on start up.

Validation ensures that the stored messages are
ExtendedChannelAnnouncementMessages and that a valid chain_client
exists to check for the existance of UTXOs. If the UTXO does
not exist, it is removed using the .removeChannel() method
on the GossipManager.

This code should round out full restore of the Gossip information
when things start.

One additional note is that the use of instanceof can have issues
if there are different version of @lntools/wire used by the
GossipStore and the GossipManager.