crypto: adds check for maximum expansion rounds allowed.
This code and test prevents bad expansion counters if the number
of expansion rounds exceeds a single byte. Therefore output length
is limited to a total of 254 expansions. The actual byte length
depends on the hash function being used.

This code can be modified in the future if larger output key lengths
are required.