wire: gossip_manager tracks last seen block
This commit fixes the gossip_manager so that it will correctly fetch missing
data from a peer when a reconnection event occurs. Prior to this commit it
would use the block height obtained during the initial sync with the peer.

This commit modifies the gossip_manager to perform two actions:

1. It now tracks the highest chan_ann messages block height
2. It corrects an issue where subsequent reconnection events triggered a sync
   based on the originating (closure scoped) block height that was obtained
   from the gossip store.