graph: modifies LndSerializer to more closely align with describeGraph
This change fixes an issue with the use of `channels` instead of `edges` and
is now inline with the LNDs describeGraph API call.

This change also modifies the serializer to include two methods for serialization
`toObject` will construct the object while `toJSON` will convert the graph object
into a string.