wire: fixing TlvValueWriter TUInt64
There was an issue where a lower byte value was being incorrectly written
due to how .toString(16) returns a value that is not a value byte representation
in hex. This would cause numbers such as 0x4d2 to be incorrectly written since
Node.js Buffers are not created with a partial byte value.  Such as
Buffer.from("4d2", "hex") ignores the "2".