wire: modify gossip filter to prevent duplicate chan_ann
This is a reversion + additional logic of prior commit
https://github.com/altangent/lntools/commit/1496e8c292c4212fb4ab799b6c8f5a42ef34d743

This change adds back the blocking of duplicate chan_ann messages that was removed
in a prior commit. Ignoring duplicate messages is a requirement to allow efficient restoration
of messages from a remote node. With dupe preventions, you can ignore certain messages and
prevent sig validation and onchain connections.

This fix is a prereq before changing the default query_channel_range message from starting at
block number 0.

This change adds a few other checks however. The prior commit listed above was added so that
chan_ann message that did not have onchain information added could be updated if the message
is seen again. So... this commit adds two checks.

1) if the prior message is an ext_cahn_ann
then we can just ignore the chan_ann message because we already have the information
that we need for that message.
2) if there is not a chain client then ignore the message because there is no
update that could possibly be made.

This logic now allows for a chan_ann to be updated to an ext_chan_ann if it was saved. But
an ext_chan_ann will always result in a skipped message.