wire: change historical gossip sync for gossip_queries to start from 0
Desc - prior to this commit, the historical gossip sync starts at the the prior restored
block height - 144 blocks (1 days worth of blocks). This change makes it always start
from position 0 so that a full sync is performed from the remote peer.

Reason - query_channel_range will return blocks within a specific range but it only gives
you the chans that were created in that range. It doesn't tell you if a chan_upd happened
after a period of time meaning you end up having out of sync information for channels becase
there are out of date chan_upd messages and no way to know if they are out of date.

With gossip_queries_ex enabled, it's also possible to retrieve the timestamps and checksums
of the chan_updates but you still need to retrieve this information for all channels.

To fix the issue of out of date chan_upd messages and to make this work better with gossip_queries_ex
queries, the block height is being changed to 0.