gossip-examples: adding gossip_manager flushed event
Desc - adding a flushed event handler and logger output when it happens

Reason - this helps diagnose and view when messages get flushed through the filter.
This was useful for checking how a full historical sync works and when it is complete.
It's also interesting information to see for how general gossip functions.