bitcoind: adding retry policy support for RPC requests
This commit adds a number of classes to support the creation and use of retry
and backoff policies when making RPC requests.

Prior to this commit, RPC failures would result in process failure and there
was no mechanism for initiating an automatic retry.

Future work should include adding logging support for the policies.