bitcoind: fixing problem with request errors that resolve
This adds a JsonRpcError type that is the rejection result when the RPC
responds with a non-200 status code. This custom error object will also include
the statusCode and body of the response.

This fixes the jsonrpcRequest method to ensure that errors are returned instead
of returning a blank result