gossip: change reply_channgel_range complete to full_information
Prior to PR #730 in BOLT7, reply_channel_range field was called complete. This
change modified the byte field to be full_information to disambiguate the
meaning. It now indicates that the full_information is available to the remote
peer. Previously complete was used for both this purpose and by some
implementations to indicate that a multipart reply was not yet complete. This
latter functionality is now handled by the sequencing of the block ranges in
the reply message.