wire: restructuring peer construction
This commit removes the peer options object and simplifies the constructor of
a peer to just include the local secret, remote secret, init factory, and
logger.

The `connect` method now includes host/port parameters.

The `attach` method was also created that allows attachment of a socket instead
of the creation of a socket used in `connect`.

These changes ensure that the constructor for a Peer has all required
information, but the initialization / attachment to the socket now occurs
via the specific `connect` or `attach` methods.