noise: capture the rpk on NoiseSocket
This change captures the resolved remote_pub_key into the socket  at the
completion of the handshake process when the socket is acting as a responder.

When the socket is acting as a responder (eg: from a server), the initiator
knows the pubkey of the server, but the server does not know the pubkey of the
initiator until the completion of the handshake. In order for the socket.rpk
to have a value, we need to capture here.

When you are the initiator we immediately capture the rpk since we already know
it.