wire: initialization should validate supported chain_hash
With the inclusion of the network TLV, a peer can validate whether it knows
anything about the chain_hashes that the remote note is connecting with.

To accomplish this, Peer and PeerServer now includes a list of supported
chain_hashes for the local node. Upon initialization, the supplied chain_hashes
are validated against the local chain_hashes. If none are supported the peer
disconnects.