gossip,peer: adding raw send method to peer
This change adds the send method to the peer and messageSender interfaces.

This method allows for sending raw buffer data to the peer and can be used
when a client wants to control the serialization of a messages. The sendMessage
method can be used to send a message using the default serialization inputs
for the message.

This fixes a need to control serialization methods as controlled by features,
not something that is currently supported, but will likely need to be supported
in the futures.