gossip: creating GossipRelay to manage rumor mongering
This change adds a basic GossipRelay class that abstracts logic for
handling rumor mongering. The original pass is a naive implementation
that maintains a single queue of gossip messages and maintains index
positions in the queue for all peers.

Messages are removed from the queue based on a maximum queue depth which
must be tuned based on the expected size of the network graph. Otherwise
messages may become hot again.