gossip: Fixing uncaught error for ChannelRangeQuery
This change modifies the way that success and failure transitions happen
for the ChannelRangeQuery state machine. Prior to this change, an
event emitter was used and there was a `once` binding to handle the first
success or error event that was fired. The challenge here is that it
doesn't limit multiple events from actually firing, only that the
promise would be resolved once.

This meant that when multiple "error" events were triggered the uncaught
error event would bubble up to the caller (at the Peer) message handler.

To resolve this, we now capture the resolve and reject methods for the
promise and make the state transitions idempotent so that multiple
inbound events won't impact the state transition.