gossip: Fixing ChannelRangeQuery and ChannelsQuery messages after final
This commit is actually the underlying cause of the duplicate failure
messages in prior commits. This change executes much needed clean-up
for the query commands so that once the state transitions to a completed
state the query commands no longer listen for messages from the peer.

This actually resolves the prior issues, but we'll leave that code in
place since the event emitters are not super efficient.