gossip: Eases detection of legacy LND reply_channel_range message
Legacy LND gossip_query implementations did not correctly use the
full_information flag. It used this flag to indicate the competion of
multipart messages. This creates an issue where full_information=false
should indicate a failure in the ability for a receiver to serve the
query.

This change loosens the check on full_information so that if we detect
short_channel_ids and the full_information=false then we use the
full_information=true as the sentinel value for completion. Otherwise
we can use the sequence of returned block numbers in accordance with
the specification.