bitcoin: creating HashValue type
Creates a non-primative type for bitcoin related hash values. Because
Bitcoin has varying uses for serialization it can be difficult to
decipher what byte ordering is being used by a hashed value (such as a
block hash, tx hash, or merkle root). This class removes that ambiguity
and can be used by Tx's, Block's and TxIn's to allow for easier
de/ser operations.