bufio: add unshift method to StreamReader
When reading from a stream, you may need to push bytes back into the
stream. An example of usage would be checking for a market byte that
can be interpreted in multiple ways.

This commit adds the `.unshift` method to `StreamReader` to support
this operation, which relies on the ReadableStream's implementation of
unshift under the hood.