chainmon: create BlockDiffer type
This type takes two block headers and finds a common ancestor as well
as connect/disconnect that need to happen to get get the last known
block into the current block.

This type will be used by the BlockWatcher to perform a resync or when
new blocks arrive to perform a reorg check.