gossip: fixes channel sync for large results
This fixes an error where a large number of scids (65k+) was resulting
in a RangeError: Maximum stack size exceeded. Upon investigation, the
ChannelQuery class takes an array of SCIDs and uses array.push(...scids).

Using push in this manner uses .apply under the covers. Apply uses the
stack.  Therefore the number of items you can push is limited by the
stack size.

The code was changed to use an iterator and push a single element at a
time.  An integration test was added to ensure there are not problems
in the future.