noise: change NoiseSocket to object mode
This sets the NoiseSocket to run in object mode and adds a highwatermark
property that can be customized.

This is performed so that individual wire messages will be streamed as a
single buffer object. If this is left in normal mode then buffers
will be contanenated when the stream is not in flowing mode.