bitcoin: refactor PublicKey to use compressed prop
This changes PublicKey so that it has new property `compressed`. This
changes the constructor to accept this readonly property. It modifies
the `to*` methods so they no longer need the compressed parameter and
instead use the internal `compressed` property. Lastly there is now
a `toPubKey(compressed: boolean)` method that allow conversion bewteen
the types.

The impetus of this change is that it simplifies a number of method
signatures and allow us to capture the compressed status of a particular
pubkey. This is important for things like WIF.