docs: fix failure results for open channel
This modifies the failure conditions for open channel states so that
reconnection failures result in an end transition to "local unilateral
close".  Prior to that any invalid states result in an "abandon" final
stage, however the effect is split between "fail" and "forget". Under
a "fail" condition, you can optionally send the error message to the
peer.