bitcoind: url support in options
This simplifies the options to allow the user to provide a url instead
of a host and port as two seperate options. This simplifies the calling
code and removes some of the complexities of node url.parse and
http.request. Additionally this change constructs a separate JSON RPC
option object so the complexity of url parsing/http requests doesn't
bleed up to the calling code.

url.parse host includes the port ie: 127.0.0.1:18443 in addition to the
raw hostname value. http.request allows the use of both host and
hostname, with hostname taking precedence over host to support usecases
where the object originates with url.parse.