gossip: refactor message event flow
Refactors the web of Peer message events into orchestrated method calls.
More work needs to be done still but this removes the "message" event
passing and the wrapping of a Peer with a GossipPeer, that subsequently
emitted message events. Finally, this changed GossipManager so that it
no longer emits events.