channels: implement initial state machine management
This implemeents a StateMachine type (could be abstracted but its not)
that implements the state pattern.  These types exist purely to
perform state transitions and mutate the `Channel` type.  The `Channel`
type maintains a reference to its current StateMachine. Finally the
ChannelManager is the interpreter of events whereby it exists to:

1. fetch a channel(s) for an event
2. execute the event on each state machine
3. transition the channel to the next state