bitcoin: addInput/Output no longer clone
This leaves cloning up to the caller. The caller can then capture a
reference to the argument and mutate it after it is part of the builder.

For example, when changing the amount, it is more common to capture a
reference of the object, add it to the builder, then perform a mutation
of the output after the fact. The existing code unexpectedly breaks this
where mutating the argument object has no effect on the TxBuilder.