lightning: refactor state machine to use events
This is part of a broader refactor to use composable state pattern
instead of a class hierarchy based state pattern. To do this we create
a transition factory that constructs transition logic and allow
the state machine type to dynamically add transitions for events.

Additional commits will remove existing state machine sub-classes.